[jit] Add a utility class SourceRef to represent Source as keys
Summary:
A new type SourceRef is introduced to represent a unique identifier to source
text. The type holds refcount to underlying source, and supports comparators
and hash functions, such that it can be used in C++ and Python maps. In later
diffs we will use this to aggregate and print profiling information.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 9fe75ba50f259f07972bc6d681142cd71f305b15
Pull Request resolved: https://github.com/pytorch/pytorch/pull/57396