Remove getScriptRemoteCallType
Because OwnerRRefs used to be created before their value was computed, we had to figure out their type ahead of time. After the previous diff, we inverted the order of operations, and we can now first compute the result and then create the OwnerRRef. Which means we can just inspect the value to get its type. Much simpler, and much less likely to get it wrong.

Differential Revision: [D28253843](https://our.internmc.facebook.com/intern/diff/D28253843/)

ghstack-source-id: 128413017
Pull Request resolved: https://github.com/pytorch/pytorch/pull/57854