Unify fetching RRefs
Just like with assigning OwnerRRefs, we can also deduplicate the code paths for fetching their values. In fact this was duplicated three times, with different ways of post-processing the value (once for JIT, once for Python, once for autograd). Thanks to future, we can have that logic once, and then connect it to different follow-up steps.

Differential Revision: [D28286172](https://our.internmc.facebook.com/intern/diff/D28286172/)

ghstack-source-id: 128418055
Pull Request resolved: https://github.com/pytorch/pytorch/pull/57859