Add support to to(device) op. (#56857)
Currently, if we call tensor.to() method and pass a device as the parameter. It will fail, because in symbolic function of to() we didn't handle such case.

So add a check in the beginning of this symbolic function, if this is a device cast, we return self directly. A test has also been added.

Co-authored-by: Jay Zhang <jiz@microsoft.com>

ghstack-source-id: a0336a8a924635e6551a6986472b6a7426290112
Pull Request resolved: https://github.com/pytorch/pytorch/pull/57599